api/tenants/requirements.txt
# Core Django Dependencies
Django>=4.0,<5.0
djangorestframework>=3.14.0,<4.0
django-filter>=22.1,<23.0
django-cors-headers>=3.13.0,<4.0
django-extensions>=3.1.0,<4.0

# Database
psycopg2-binary>=2.9.0,<3.0

# Cache and Session
redis>=4.3.0,<5.0
django-redis>=5.2.0,<6.0

# Background Tasks
celery>=5.2.0,<6.0
django-celery-beat>=2.4.0,<3.0
django-celery-results>=2.4.0,<3.0

# Security and Authentication
cryptography>=3.4.0,<4.0
python-jose>=3.3.0,<4.0
passlib>=1.7.0,<2.0

# File Storage
django-storages>=1.13.0,<2.0
boto3>=1.26.0,<2.0  # For S3 storage

# Email
django-anymail>=2.0.0,<3.0
sendgrid>=6.9.0,<7.0

# API and Serialization
drf-spectacular>=0.24.0,<1.0
drf-spectacular-sidecar>=0.1.0,<1.0

# Date and Time
python-dateutil>=2.8.0,<3.0
pytz>=2022.1,<2023.0

# Validation
marshmallow>=3.14.0,<4.0
jsonschema>=4.0.0,<5.0

# HTTP Client
requests>=2.28.0,<3.0
urllib3>=1.26.0,<2.0

# Image Processing
Pillow>=9.0.0,<10.0

# CSV and Excel
openpyxl>=3.0.0,<4.0
pandas>=1.5.0,<2.0

# Monitoring and Logging
sentry-sdk>=1.9.0,<2.0
django-debug-toolbar>=3.8.0,<4.0

# Testing
pytest>=7.2.0,<8.0
pytest-django>=4.5.0,<5.0
pytest-cov>=4.0.0,<5.0
factory-boy>=3.2.0,<4.0
faker>=15.0.0,<16.0

# Development
black>=22.0.0,<23.0
flake8>=5.0.0,<6.0
isort>=5.10.0,<6.0
mypy>=0.991,<1.0

# Environment Configuration
python-dotenv>=0.19.0,<1.0

# Webhook Support
webhooks>=0.4.0,<1.0

# Analytics
django-silk>=4.2.0,<5.0

# Rate Limiting
django-ratelimit>=3.5.0,<4.0

# Pagination
django-paginator>=2.4.0,<3.0

# File Upload
django-file-upload>=1.0.0,<2.0

# Internationalization
django-modeltranslation>=0.18.0,<19.0

# Background Task Monitoring
flower>=1.2.0,<2.0

# API Documentation
drf-yasg>=1.21.0,<2.0
coreapi>=2.3.0,<3.0

# Email Templates
django-templated-mail>=1.4.0,<2.0

# Validation
django-rest-framework-serializer-extensions>=0.7.0,<1.0

# Performance
django-debug-toolbar-silk>=0.8.0,<1.0

# Security
django-axes>=5.40.0,<6.0
django-otp>=1.0.0,<2.0

# Background File Processing
django-background-tasks>=1.2.0,<2.0

# Timezone Support
django-timezone-field>=4.2.0,<5.0

# API Versioning
django-rest-framework-versioning>=1.0.0,<2.0

# Swagger/OpenAPI
drf-openapi>=1.2.0,<2.0

# Caching
django-cacheops>=6.0.0,<7.0

# Database Migrations
django-migration-extensions>=0.3.0,<1.0

# Email Queue
django-mailer>=2.3.0,<3.0

# Health Checks
django-health-check>=3.16.0,<4.0

# Configuration Management
django-environ>=0.9.0,<1.0

# Logging
django-log-request-id>=2.1.0,<3.0

# API Rate Limiting
drf-extra-fields>=3.4.0,<4.0

# Background Task Scheduling
django-q>=1.3.0,<2.0

# File Compression
django-compressor>=4.1.0,<5.0

# Database Utilities
django-db-connection-pool>=1.2.0,<2.0

# API Testing
django-test-plus>=1.4.0,<2.0

# Model Utilities
django-model-utils>=4.2.0,<5.0

# JSON Field Support
django-jsonfield>=3.1.0,<4.0

# UUID Field Support
django-extensions>=3.1.0,<4.0

# Time Utilities
django-timezone-utils>=0.3.0,<1.0

# Email Validation
django-email-verification>=1.0.0,<2.0
